🛡️ Sentinel: Add input validation for drawdown plan service
